feat: use types
